requirements.txt
ansible_lint==6.14.4
openai==0.27.2
PyYAML==5.4.1
PyYAML==6.0
Requests==2.30.0
yamllint==1.30.0
